+++ flows/sheet_delete.py
@@ -1,12 +1,13 @@
 from config import settings
 
 from .base_flow import BaseLinearFlow
-from apis.irods_utils import get_project_path,\
-    get_project_group_name
+from apis.irods_utils import get_project_path
 from tasks import omics_tasks, irods_tasks
 
 
 PROJECT_ROOT = settings.TASKFLOW_IRODS_PROJECT_ROOT
+TASKFLOW_SAMPLE_DIR = settings.TASKFLOW_SAMPLE_DIR
+TASKFLOW_LANDING_ZONE_DIR = settings.TASKFLOW_LANDING_ZONE_DIR
 
 
 class Flow(BaseLinearFlow):
@@ -23,11 +24,19 @@ class Flow(BaseLinearFlow):
         ########
 
         project_path = get_project_path(self.project_uuid)
-        sample_path = project_path + '/bio_samples'
+        sample_path = project_path + '/' + TASKFLOW_SAMPLE_DIR
+        zone_path = project_path + '/' + TASKFLOW_LANDING_ZONE_DIR
 
         ##############
         # iRODS Tasks
         ##############
+
+        self.add_task(
+            irods_tasks.RemoveCollectionTask(
+                name='Remove sample sheet landing zones',
+                irods=self.irods,
+                inject={
+                    'path': zone_path}))
 
         self.add_task(
             irods_tasks.RemoveCollectionTask(